+++ src/main/java/subway/path/application/MinimumTimePathStrategy.java
@@ -4,27 +4,50 @@ import org.jgrapht.graph.DefaultWeightedEdge;
 import org.jgrapht.graph.WeightedMultigraph;
 import org.springframework.stereotype.Component;
 import subway.line.domain.Section;
+import subway.path.application.dto.PathRetrieveResponse;
+import subway.station.application.dto.StationResponse;
 import subway.station.domain.Station;
 
 import java.util.List;
 
 @Component
-public class MinimumTimePathStrategy implements PathStrategy { // TODO
+public class MinimumTimePathStrategy extends AbstractPathStrategy implements PathStrategy { // TODO
+
     @Override
-    public Double getWeightOfPath(WeightedMultigraph<Station, DefaultWeightedEdge> graph,
-                                  Station sourceStation,
-                                  Station targetStation) {
-        // MinimumTimePathFinder에서의 getWeightOfMinimumTimePath() 메서드의 구현을 여기에 넣습니다.
+    public PathRetrieveResponse findPath(WeightedMultigraph<Station, DefaultWeightedEdge> graph, List<Section> sections, List<Station> stationsInPath, Station sourceStation, Station targetStation) {
+//        validIsSameOriginStation(sourceStation, targetStation);
+//
+//        List<Station> stations = getStations(sections);
+//        WeightedMultigraph<Station, DefaultWeightedEdge> graph = getGraph(sections, stations);
+//
+//        List<Station> stationsInPath = getPath(graph, sourceStation, targetStation);
+        Long totalDistance = getDistanceInMinimumTimePath(stationsInPath, sections);
+        Double minimumWeight = getWeightOfPath(graph, sourceStation, targetStation);
+
+        return PathRetrieveResponse.builder()
+                .stations(StationResponse.from(stationsInPath))
+                .distance(totalDistance)
+                .duration(minimumWeight.longValue())
+                .build();
     }
 
     @Override
     public Long getTotalWeightInPath(List<Station> stationsInPath, List<Section> sections) {
-        // MinimumTimePathFinder에서의 getDistanceInMinimumTimePath() 메서드의 구현을 여기에 넣습니다.
+        List<Section> sectionsInPath = getSections(stationsInPath, sections);
+        return sectionsInPath.stream()
+                .map(Section::getDistance)
+                .reduce(0L, Long::sum);
     }
 
     @Override
     public void setEdgeWeight(WeightedMultigraph<Station, DefaultWeightedEdge> graph, Section section, DefaultWeightedEdge edge) {
-        // edge의 weight를 섹션의 duration으로 설정
         graph.setEdgeWeight(edge, section.getDuration());
+    }
+
+    private Long getDistanceInMinimumTimePath(List<Station> stationsInPath, List<Section> sections) {
+        List<Section> sectionsInPath = getSections(stationsInPath, sections);
+        return sectionsInPath.stream()
+                .map(Section::getDistance)
+                .reduce(0L, Long::sum);
     }
 }
