+++ src/main/java/subway/path/application/PathFinder.java
@@ -1,29 +1,58 @@
 package subway.path.application;
 
+import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.DefaultWeightedEdge;
 import org.jgrapht.graph.WeightedMultigraph;
-import org.springframework.stereotype.Component;
+import subway.constant.SubwayMessage;
+import subway.exception.SubwayBadRequestException;
 import subway.line.domain.Section;
 import subway.path.application.dto.PathRetrieveResponse;
 import subway.station.domain.Station;
 
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class PathFinder {  // TODO
     private final PathStrategy strategy;
 
-    // strategy를 constructor에서 주입
     public PathFinder(PathStrategy strategy) {
         this.strategy = strategy;
     }
 
-    public PathRetrieveResponse findPath(List<Section> sections,
-                                         Station sourceStation,
-                                         Station targetStation) {
-        return strategy.findPath(sections, sourceStation, targetStation);
+    public PathRetrieveResponse findPath(List<Section> sections, Station sourceStation, Station targetStation) {
+        validIsSameOriginStation(sourceStation, targetStation);
+
+        List<Station> stations = getStations(sections);
+        WeightedMultigraph<Station, DefaultWeightedEdge> graph = getGraph(sections, stations);
+
+        List<Station> stationsIntPath = getPath(graph, sourceStation, targetStation);
+
+        return strategy.findPath(graph, sections, stationsIntPath, sourceStation, targetStation);
     }
 
-    // TODO: 나머지 메서드들...
+    private void validIsSameOriginStation(Station sourceStation, Station targetStation) {
+        if (sourceStation.equals(targetStation)) {
+            throw new SubwayBadRequestException(SubwayMessage.PATH_REQUEST_STATION_IS_SAME_ORIGIN);
+        }
+    }
+
+    private List<Station> getPath(WeightedMultigraph<Station, DefaultWeightedEdge> graph,
+                                  Station sourceStation,
+                                  Station targetStation) {
+        try {
+            DijkstraShortestPath<Station, DefaultWeightedEdge> dijkstraShortestPath = new DijkstraShortestPath<>(graph);
+            return dijkstraShortestPath.getPath(sourceStation, targetStation).getVertexList();
+        } catch (IllegalArgumentException | NullPointerException e) {
+            throw new SubwayBadRequestException(SubwayMessage.PATH_NOT_CONNECTED_IN_SECTION);
+        }
+    }
+
+    private List<Station> getStations(List<Section> sections) {
+        return sections.stream()
+                .flatMap(section -> Stream.of(section.getUpStation(), section.getDownStation()))
+                .collect(Collectors.toList());
+    }
 
     private WeightedMultigraph<Station, DefaultWeightedEdge> getGraph(List<Section> sections, List<Station> stations) {
         WeightedMultigraph<Station, DefaultWeightedEdge> graph = new WeightedMultigraph<>(DefaultWeightedEdge.class);
