+++ src/test/java/subway/unit/path/ShortestDistancePathStrategyTest.java
@@ -7,7 +7,9 @@ import subway.acceptance.station.StationFixture;
 import subway.exception.SubwayBadRequestException;
 import subway.line.domain.Line;
 import subway.line.domain.Section;
-import subway.path.application.MinimumTimePathFinder;
+import subway.path.application.PathFinder;
+import subway.path.application.ShortestDistancePathFinder;
+import subway.path.application.ShortestDistancePathStrategy;
 import subway.path.application.dto.PathRetrieveResponse;
 import subway.station.application.dto.StationResponse;
 import subway.station.domain.Station;
@@ -19,9 +21,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static subway.acceptance.station.StationFixture.getStation;
 
-@DisplayName("MinimumTimePathFinder 단위 테스트")
-public class MinimumTimePathFinderTest {
-    private MinimumTimePathFinder minimumTimePathFinder;
+@DisplayName("ShortestDistancePathFinder 단위 테스트")
+public class ShortestDistancePathStrategyTest {
+    private PathFinder shortestDistancePathFinder;
 
     private Line 이호선;
     private Line 삼호선;
@@ -45,7 +47,8 @@ public class MinimumTimePathFinderTest {
 
     @BeforeEach
     void beforeEach() {
-        minimumTimePathFinder = new MinimumTimePathFinder();
+        ShortestDistancePathStrategy shortestDistancePathStrategy = new ShortestDistancePathStrategy();
+        shortestDistancePathFinder = new PathFinder(shortestDistancePathStrategy);
 
         StationFixture.기본_역_생성();
 
@@ -70,25 +73,26 @@ public class MinimumTimePathFinderTest {
 
     }
 
+
     /**
      * Given 구간이 있을 때
      * When 경로 조회를 하면
      * Then 경로 내 역이 목록으로 반환된다
-     * Then 경로의 총 시간이 반환된다
+     * Then 경로의 총 길이가 반환된다
      */
-    @DisplayName("최소시간 경로 조회")
+    @DisplayName("최단거리 경로 조회")
     @Test
-    void getMinimumTimePath() {
+    void getShortestDistancePath() {
         // when
-        PathRetrieveResponse shortestPath = minimumTimePathFinder.findPath(구간목록, getStation("교대역"), getStation("양재역"));
+        PathRetrieveResponse shortestPath = shortestDistancePathFinder.findPath(구간목록, getStation("교대역"), getStation("양재역"));
 
         // then
         assertThat(shortestPath.getStations())
                 .extracting(StationResponse::getName)
-                .containsExactlyInAnyOrder("교대역", "강남역", "양재역");
+                .containsExactlyInAnyOrder("교대역", "남부터미널역", "양재역");
 
         // then
-        assertThat(shortestPath.getDuration()).isEqualTo(11L);
+        assertThat(shortestPath.getDistance()).isEqualTo(5L);
     }
 
     /**
@@ -96,11 +100,11 @@ public class MinimumTimePathFinderTest {
      * When 동일한 구간을 조회하면
      * Then 조회 되지 않는다.
      */
-    @DisplayName("최소시간 경로 조회 : 요청 구간 동일")
+    @DisplayName("최단거리 경로 조회 : 요청 구간 동일")
     @Test
-    void getMinimumTimePathWithSameOrigin() {
+    void getShortestDistancePathWithSameOrigin() {
         // when/then
-        assertThatThrownBy(() -> minimumTimePathFinder.findPath(구간목록, getStation("교대역"), getStation("교대역")))
+        assertThatThrownBy(() -> shortestDistancePathFinder.findPath(구간목록, getStation("교대역"), getStation("교대역")))
                 .isInstanceOf(SubwayBadRequestException.class);
     }
 
@@ -109,11 +113,11 @@ public class MinimumTimePathFinderTest {
      * When 연결되지 않은 구간을 조회하면
      * Then 조회 되지 않는다.
      */
-    @DisplayName("최소시간 경로 조회 : 연결되지 않은 구간")
+    @DisplayName("최단거리 경로 조회 : 연결되지 않은 구간")
     @Test
-    void getMinimumTimePathNotConnectedSection() {
+    void getShortestDistancePathNotConnectedSection() {
         // when/then
-        assertThatThrownBy(() -> minimumTimePathFinder.findPath(구간목록, getStation("교대역"), getStation("왕십리역")))
+        assertThatThrownBy(() -> shortestDistancePathFinder.findPath(구간목록, getStation("교대역"), getStation("왕십리역")))
                 .isInstanceOf(SubwayBadRequestException.class);
     }
 
@@ -122,15 +126,12 @@ public class MinimumTimePathFinderTest {
      * When 존재하지 않는 역을 조회하면
      * Then 조회 되지 않는다.
      */
-    @DisplayName("최소시간 경로 조회 : 존재하지 않는 역")
+    @DisplayName("최단거리 경로 조회 : 존재하지 않는 역")
     @Test
-    void getMinimumTimePathNotExistStation() {
+    void getShortestDistancePathNotExistStation() {
         // when/then
-        assertThatThrownBy(() -> minimumTimePathFinder.findPath(구간목록, new Station(99L,"그런역"),  new Station(98L,"저런역")))
+        assertThatThrownBy(() -> shortestDistancePathFinder.findPath(구간목록, new Station(99L, "그런역"), new Station(98L, "저런역")))
                 .isInstanceOf(SubwayBadRequestException.class);
     }
-
-
-
 
 }
